Refactoring name variable for id, add class for validation local variable!!!

+++ src/main/java/ru/yandex/practicum/filmorate/validation/impl/NotNegativeValueValidator.java
@@ -7,10 +7,10 @@ import jakarta.validation.constraintvalidation.ValidationTarget;
 import ru.yandex.practicum.filmorate.validation.NotNegativeValue;
 
 @SupportedValidationTarget(ValidationTarget.ANNOTATED_ELEMENT)
-public class NotNegativeValueValidator implements ConstraintValidator<NotNegativeValue, Long> {
+public class NotNegativeValueValidator implements ConstraintValidator<NotNegativeValue, Integer> {
 
 	@Override
-	public boolean isValid(Long value, ConstraintValidatorContext context) {
+	public boolean isValid(Integer value, ConstraintValidatorContext context) {
 		if (value == null)
 			return true;
 		return value > 0;
